--- Assets/PlayerController.cs
@@ -1,29 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerController : MonoBehaviour
-{
-
-    [SerializeField]
-    public Vector2 speed = new Vector2(50, 50);
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        float inputX = Input.GetAxis("Horizontal");
-        float inputY = Input.GetAxis("Vertical");
-
-        Vector3 move = new Vector3(speed.x * inputX, speed.y * inputY, 0);
-
-        move *= Time.deltaTime;
-
-        transform.Translate(move);
-    }
-}
