+++ Assets/README.txt
@@ -1,0 +1,1 @@
+La scena si trova in Scenes/SampleScene
