+++ Assets/README.txt.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4a723e92df498fb4e8c86a5f324bf663
-DefaultImporter:
+guid: 8b6f5d160145fba4ab3630d13169be39
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
