+++ Assets/Scripts.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 4a723e92df498fb4e8c86a5f324bf663
+guid: 7e0bfaa925a053646bf73d977c5953d2
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
