+++ Assets/Scripts/Player.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 4a723e92df498fb4e8c86a5f324bf663
+guid: 94cdf589ee26a8e4493e3a9b130384ca
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
